django book store application with jwt token

+++ BookStoreApp/book/views.py
@@ -1,14 +1,12 @@
+import logging
 from .serializers import BookSerializer
 from .models import Book
 from rest_framework.response import Response
 from rest_framework.views import APIView
-
 from user.models import User
 from user.serializers import UserSerializer
 from rest_framework import status, views
-import logging
-from rest_framework.exceptions import ValidationError
-
+from user.util import VerifyToken
 
 
 logger = logging.getLogger('django')
@@ -24,14 +22,17 @@ class BookView(APIView):
         :return: book stored successfully or not
         """
         try:
-            serializer = BookSerializer(data=request.data)
-            serializer.is_valid(raise_exception=True)
-            serializer.save()
-            return Response({"message": "book store successfully", "data": serializer.data},
-                            status=status.HTTP_201_CREATED)
+            if VerifyToken().verify_token(request):
+                serializer = BookSerializer(data=request.data)
+                serializer.is_valid(raise_exception=True)
+                serializer.save()
+                return Response({"message": "book store successfully", "data": serializer.data},
+                                status=status.HTTP_201_CREATED)
+            return Response({"message": "book store unsuccessful"},
+                                status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             logging.error(e)
-            return Response({"message": "validation failed"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"message": "book with title already exists."}, status=status.HTTP_400_BAD_REQUEST)
 
 
     def get(self, request):
@@ -39,22 +40,23 @@ class BookView(APIView):
         :return: get all books
         """
         try:
-            book = Book.objects.all()
-            serializer = BookSerializer(book, many=True)
-            return Response(
-                {
-                    "message": "Here your Book",
-                    "data": serializer.data
-                },
-                status=status.HTTP_201_CREATED)
+            if VerifyToken().verify_token(request):
+                book = Book.objects.all()
+                serializer = BookSerializer(book, many=True)
+                return Response(
+                    {
+                        "message": "Here your Book",
+                        "data": serializer.data
+                    },
+                    status=status.HTTP_201_CREATED)
         except Exception as e:
-            print(e)
             logging.error(e)
             return Response(
                 {
                     "message": "No book for you"
                 },
                 status=status.HTTP_400_BAD_REQUEST)
+
 
     def put(self, request):
         """
@@ -63,25 +65,25 @@ class BookView(APIView):
         :return: Response
         """
         try:
-            print(request.data["title"])
-            book = Book.objects.get(title=request.data["title"])
-            serializer = BookSerializer(book, data=request.data)
-            serializer.is_valid(raise_exception=True)
-            serializer.save()
-            return Response(
-                {
-                    "message": "Data updated successfully",
-                    "data": serializer.data
-                },
-                status=status.HTTP_201_CREATED)
+            if VerifyToken().verify_token(request):
+                book = Book.objects.get(title=request.data["title"])
+                serializer = BookSerializer(book, data=request.data)
+                serializer.is_valid(raise_exception=True)
+                serializer.save()
+                return Response(
+                    {
+                        "message": "Data updated successfully",
+                        "data": serializer.data
+                    },
+                    status=status.HTTP_201_CREATED)
         except Exception as e:
-            print(e)
             logging.error(e)
             return Response(
                 {
                     "message": "Data not updated"
                 },
                 status=status.HTTP_400_BAD_REQUEST)
+
 
     def delete(self, request):
         """
@@ -90,13 +92,14 @@ class BookView(APIView):
         :return: response
         """
         try:
-            book = Book.objects.get(title=request.data["title"])
-            book.delete()
-            return Response(
-                {
-                    "message": "Data deleted"
-                },
-                status=status.HTTP_204_NO_CONTENT)
+            if VerifyToken().verify_token(request):
+                book = Book.objects.get(title=request.data["title"])
+                book.delete()
+                return Response(
+                    {
+                        "message": "Data deleted"
+                    },
+                    status=status.HTTP_204_NO_CONTENT)
         except Exception as e:
             logging.error(e)
             return Response(
@@ -104,4 +107,3 @@ class BookView(APIView):
                     "message": "Data not deleted"
                 },
                 status=status.HTTP_400_BAD_REQUEST)
-
